BugID:17946655: link security v0.5:
1. support both customized and original ssl/tls.
2. alicrypto code reconstruction.
3. uniform security header file path.
4. add dpm module.
5. xx.mk -> aos.mk support aos comp.

Reviewed-on: http://yuncode.alibaba-inc.com/177493
Tested-by: 李诚 lc122798 <lc122798@alibaba-inc.com>

+++ security/isst/src/wrapper/km/sst_wrapper.c
@@ -145,7 +145,7 @@ static uint32_t _sst_enc_data(void *p_sst, const uint8_t *data_in)
     /* enc data */
     ret = km_envelope_finish(envelope_ctx, (uint8_t *)data_in, data_len, p_data, &data_len);
     if (ret || data_len != p_sst_obj->data_size) {
-        SST_ERR("envelope finish failed 0x%x\n", ret);
+        SST_ERR("envelope finish failed 0x%x\n", (unsigned int)ret);
         return SST_ERROR_GENERIC;
     }
 
@@ -174,7 +174,7 @@ static uint32_t _sst_dec_data(void *p_sst, uint8_t *data_out)
 
     /* decrypt get subkey*/
     if (p_sst_obj->key_size != 16) {
-        SST_ERR("key size error key_size = %d\n", p_sst_obj->key_size);
+        SST_ERR("key size error key_size = %d\n", (unsigned int)p_sst_obj->key_size);
         return SST_ERROR_GENERIC;
     }
